+++ src/components/Rockets.jsx
@@ -1,7 +1,8 @@
 import React, { useEffect } from 'react';
 import '../style/rockets.css';
 import { useSelector, useDispatch } from 'react-redux';
-import { getRocketData, reserveRocket } from '../redux/rockets/rocketsSlice';
+import { getRocketData } from '../redux/rockets/rocketsSlice';
+import Rocket from './RocketCard';
 
 function Rockets() {
   const { rockets, isLoading, error } = useSelector((store) => store.rockets);
@@ -10,10 +11,6 @@ function Rockets() {
   useEffect(() => {
     dispatch(getRocketData());
   }, [dispatch]);
-
-  const handleButtonClick = (rocketId) => {
-    dispatch(reserveRocket(rocketId));
-  };
 
   if (isLoading) {
     return <div className="alert-box">Loading data from API...</div>;
@@ -27,35 +24,7 @@ function Rockets() {
     <div className="container">
       <ul>
         {rockets.map((rocket) => (
-          <li key={rocket.id} className="rocket-item">
-            <div className="rocket-image-container">
-              <img src={rocket.imageRocket} alt="rocket_image" className="rocket-image" key={rocket.id} />
-            </div>
-            <div className="rocket-description-container">
-              <h3 className="rocket-name">{rocket.rocketName}</h3>
-              <div>
-                {rocket.isReserved && <samp>Reserved </samp>}
-                <p className="rocket-description">{rocket.description}</p>
-              </div>
-              {rocket.isReserved ? (
-                <button
-                  type="button"
-                  className="cancel-btn"
-                  onClick={() => handleButtonClick(rocket.id)}
-                >
-                  Cancel Reservation
-                </button>
-              ) : (
-                <button
-                  type="button"
-                  className="booking-btn"
-                  onClick={() => handleButtonClick(rocket.id)}
-                >
-                  BOOK ROCKET
-                </button>
-              )}
-            </div>
-          </li>
+          <Rocket key={rocket.id} rocket={rocket} />
         ))}
       </ul>
     </div>
